Fix splits and includes
