+++ config/GMSJ01/splits.txt
@@ -345,16 +345,16 @@ JSystem/J2D/J2DOrthoGraph.cpp:
 	.data       start:0x803A9BB0 end:0x803A9BD8
 	.sdata2     start:0x8040B710 end:0x8040B730
 
-JSystem/J3D/J3DVertex.cpp:
+JSystem/J3D/J3DGraphBase/J3DVertex.cpp:
 	.text       start:0x80037180 end:0x800373E8
 	.data       start:0x803A9BD8 end:0x803A9BE8
 
-JSystem/J3D/J3DPacket.cpp:
+JSystem/J3D/J3DGraphBase/J3DPacket.cpp:
 	.text       start:0x800373E8 end:0x80037A58
 	.data       start:0x803A9BE8 end:0x803A9C60
 	.sbss       start:0x80409950 end:0x80409958
 
-JSystem/J3D/J3DTevs.cpp:
+JSystem/J3D/J3DGraphBase/J3DTevs.cpp:
 	.text       start:0x80037A58 end:0x80039130
 	.rodata     start:0x80370CE0 end:0x80370D50
 	.data       start:0x803A9C60 end:0x803A9DD8
@@ -362,14 +362,14 @@ JSystem/J3D/J3DTevs.cpp:
 	.sdata      start:0x80408AF8 end:0x80408B78
 	.sdata2     start:0x8040B730 end:0x8040B748
 
-JSystem/J3D/J3DDrawBuffer.cpp:
+JSystem/J3D/J3DGraphBase/J3DDrawBuffer.cpp:
 	.text       start:0x80039130 end:0x80039858
 	.ctors      start:0x8036FBD4 end:0x8036FBD8
 	.data       start:0x803A9DD8 end:0x803A9E98
 	.sbss       start:0x80409958 end:0x80409960
 	.sdata2     start:0x8040B748 end:0x8040B758
 
-JSystem/J3D/J3DMaterialAnm.cpp:
+JSystem/J3D/J3DGraphAnimator/J3DMaterialAnm.cpp:
 	.text       start:0x80039858 end:0x80039B38
 	.data       start:0x803A9E98 end:0x803A9EB8
 
@@ -889,7 +889,7 @@ JSystem/JAudio/JASystem/JASTrackPort.cpp:
 JSystem/random.cpp:
 	.text       start:0x80068B1C end:0x80068C18
 
-JSystem/J3D/J3DMaterialFactory_v21.cpp:
+JSystem/J3D/J3DGraphLoader/J3DMaterialFactory_v21.cpp:
 	.text       start:0x80068C18 end:0x8006A1D0
 	.sdata2     start:0x8040BBC8 end:0x8040BBD8
 
@@ -900,7 +900,7 @@ JSystem/JAudio/JASystem/JASSeqParser.cpp:
 	.data       start:0x803ABEA0 end:0x803AC420
 	.sdata2     start:0x8040BBD8 end:0x8040BC18
 
-JSystem/J3D/J3DMaterialAttach.cpp:
+JSystem/J3D/J3DGraphAnimator/J3DMaterialAttach.cpp:
 	.text       start:0x8006C4E0 end:0x8006C570
 	.data       start:0x803AC420 end:0x803AC430
