+++ src/JSystem/J3D/J3DGraphAnimator/J3DModel.cpp
@@ -1,3 +1,3 @@
-#include <JSystem/J3D/J3DModel.hpp>
+#include <JSystem/J3D/J3DGraphAnimator/J3DModel.hpp>
 
 void J3DModel::entryModelData(J3DModelData*, u32, u32) { }
