+++ src/JSystem/J3D/J3DGraphBase/J3DMaterial.cpp
@@ -1,1 +1,1 @@
-#include <JSystem/J3D/J3DMaterial.hpp>
+#include <JSystem/J3D/J3DGraphBase/J3DMaterial.hpp>
